got bored, all rooms set up for basic functionality
Set up all the difficulty buttons on the main menu, all of which go to the respective difficulty room. All the rooms have columns of 5,4, and 3 chips and have a functional exit button to get back to the main menu. Chip object has a color value (may just need to make 3 different objects for each color?)